Move HTML process to an individual file
